changed error handler since a better error is already generated
My problem with was Protected and Restricted error was sent as HTTP 500
server error which I think is not a descriptive code for this. It is more
like a client error since the client tried to delete something they shouldn't have.
so I created two new APIExceptions that will replace the errors that were thrown before.